Refactor SimpleProject and ProxyAdminProject
